autentication

+++ src/controllers/ShopController.ts
@@ -8,6 +8,9 @@ export class ShopController {
   static createShop = async (req: Request, res: Response) => {
     const shop = new Shop(req.body);
 
+    // ASIGNANDO UN MANAGER
+    shop.manager = req.user.id;
+
     try {
       await shop.save();
       res.send("Compra creada correctamente");
@@ -20,7 +23,10 @@ export class ShopController {
 
   static getAllShops = async (req: Request, res: Response) => {
     try {
-      const allShops = await Shop.find({});
+      const allShops = await Shop.find({
+        $or: [{ manager: { $in: req.user.id } }],
+      });
+
       res.json(allShops);
     } catch (error) {
       console.log(error);
@@ -37,6 +43,13 @@ export class ShopController {
         const error = new Error("Compra no encontrada");
         return res.status(404).json({ error: error.message });
       }
+
+      if (shop.manager.toString() !== req.user.id.toString()) {
+        const error = new Error(
+          "No tiene los privilegios para ver esta compra"
+        );
+        return res.status(404).json({ error: error.message });
+      }
       res.json(shop);
     } catch (error) {
       console.log(error);
@@ -51,6 +64,13 @@ export class ShopController {
       const shop = await Shop.findById(id);
       if (!shop) {
         const error = new Error("Compra no encontrada");
+        return res.status(404).json({ error: error.message });
+      }
+
+      if (shop.manager.toString() !== req.user.id.toString()) {
+        const error = new Error(
+          "No tiene los privilegios para editar esta compra"
+        );
         return res.status(404).json({ error: error.message });
       }
 
@@ -75,6 +95,14 @@ export class ShopController {
         const error = new Error("Compra no encontrada");
         return res.status(404).json({ error: error.message });
       }
+
+      if (shop.manager.toString() !== req.user.id.toString()) {
+        const error = new Error(
+          "No tiene los privilegios para eliminar esta compra"
+        );
+        return res.status(404).json({ error: error.message });
+      }
+
       await shop.deleteOne();
       res.send("Compra eliminada");
     } catch (error) {
